[refactor]: 타입 정의 파일 구조화
- types/ 폴더 구조 생성 및 타입 파일 분리
   - common/: 공통 타입
   - components/: 컴포넌트 관련 타입
   - store/: 스토어 관련 타입
   - api/: API 관련 타입
