+++ src/components/Editor/EditorToolModal/LineModal/LineModal.tsx
@@ -1,6 +1,7 @@
 import * as S from "./LineModal.style";
 import useEditorStore from "@/store/useEditorStore";
-import { LineData, Line } from "./LineData";
+import { LineData } from "./LineData";
+import { Line } from "@/types/components/line.types";
 
 interface LineIconProps {
   addBlock: (type: string, data: object) => void;
