+++ src/store/useTitleStore.ts
@@ -1,44 +1,11 @@
+import { TitleStore } from "@/types/store/title.types";
 import { create } from "zustand";
-
-interface StoreProps {
-  titleText: string;
-  subtitleText: string;
-  titleCoverImage: string | null;
-  titleCoverColor: string | null;
-  isExpanded: boolean;
-  alignment: "bottom" | "center";
-  titleFont:
-    | "Noto Sans"
-    | "Nanum Myeongjo"
-    | "Nanum Gothic"
-    | "Nanum Barun Gothic"
-    | "Helvetica"
-    | "Georgia";
-  titleColor: string | null;
-
-  setTitleText: (text: string) => void;
-  setSubtitleText: (text: string) => void;
-  setTitleCoverImage: (image: string | null) => void;
-  setTitleCoverColor: (color: string | null) => void;
-  setIsExpanded: (expanded: boolean) => void;
-  setAlignment: (alignment: "bottom" | "center") => void;
-  setTitleFont: (
-    font:
-      | "Noto Sans"
-      | "Nanum Myeongjo"
-      | "Nanum Gothic"
-      | "Nanum Barun Gothic"
-      | "Helvetica"
-      | "Georgia"
-  ) => void;
-  setTitleColor: (color: string | null) => void;
-}
 
 /**
  * 글의 제목 섹션과 관련된 전역 상태를 관리하는 스토어
  * 제목, 부제목, 커버 이미지/컬러, 폰트 등의 상태와 관련 설정 함수들을 포함
  */
-const useTitleStore = create<StoreProps>((set) => ({
+const useTitleStore = create<TitleStore>((set) => ({
   titleText: "",
   subtitleText: "",
   titleCoverImage: null,
