--- src/types/Font.ts
@@ -1,7 +1,0 @@
-export type FontType =
-  | "NotoSans"
-  | "NanumMyeongjo"
-  | "NanumGothic"
-  | "NanumBarunGothic"
-  | "Helvetica"
-  | "Georgia";
