+++ src/types/api/kakao.types.ts
@@ -1,0 +1,13 @@
+export interface KakaoAddressData {
+  id: string;
+  place_name: string;
+  road_address_name: string;
+  place_url: string;
+}
+
+export interface PlaceSearchResult {
+  id: string;
+  name: string;
+  address: string;
+  url: string;
+}
