+++ src/types/components/line.types.ts
@@ -1,0 +1,5 @@
+export type Line = {
+  modalImagePosition: string;
+  inBlockImageURL: string;
+  inBlockImagePosition: string;
+};
